+++ src/Commands/Moderation/Kick.ts
@@ -27,29 +27,37 @@ export default class extends Command {
             return message.channel.send(Embed.missing_args.call(this, 1));
         }
 
-        let member: GuildMember = message.mentions.members.first();
-        if(args.length > 0 && !member) {
+        const mentions = message.mentions.members;
+        let member: GuildMember;
+        // only 1 member mentioned and the mention isn't the bot itself.
+        if(
+            (mentions.size === 1 && mentions.first().id === message.guild.me.id)
+            || (mentions.size === 0 && !isNaN(+args[0]) && args[0].length >= 17 && args[0].length <= 19)
+        ) {
             try {
-                member = await message.guild.members.fetch(args[0])
-            } catch {
-                return message.channel.send(Embed.fail(`
-                *${args[0]}* is not a valid member!
-
-                Examples:
-                \`\`kick @user for trolling\`\`
-                \`\`kick 1234567891234567\`\`
-                `));
+                member = await message.guild.members.fetch(args[0]);
+            } catch(e) {
+                this.logger.log(e.toString());
+                return message.channel.send(Embed.fail('Member couldn\'t be fetched!'));
             }
+        } else if(mentions.size > 1) { // @KhafraBot#0001 kick @user#0001 blah - @Mod#0001
+            member = mentions.first().id === message.guild.me.id ? mentions.first(2).pop() : mentions.first();
         }
 
-        if(!member.kickable) {
+        if(!(member instanceof GuildMember)) {
+            return message.channel.send(Embed.fail('Member couldn\'t be fetched!'));
+        } else if(!member.kickable) {
             return message.channel.send(Embed.fail(`${member} is not kickable!`));
         }
 
-        await member.kick(args.slice(1).join(' '));
-        
-        const embed = Embed.success(`**Successfully** kicked ${member}${args.slice(1).join(' ').length ? ' for ' + args.slice(1).join(' ') : ''}!`);
-
-        return message.channel.send(embed);
+        try {
+            await member.kick(args.slice(1).join(' '));
+            return message.channel.send(Embed.success(`Successfully kicked ${member}!`));
+        } catch(e) {
+            return message.channel.send(Embed.fail(`
+            An unexpected error occurred!
+            \`\`${e}\`\`
+            `));
+        }
     }
 }